Fix flaky unit test
failing spec: ./spec/models/custom_field/calculated_value_spec.rb:626
failing job: https://github.com/opf/openproject/actions/runs/17242723797/job/48924155525

Error was `expected ["contains custom fields that are not allowed: float, int."]
to include "contains custom fields that are not allowed: int, float."`.

The error message is not deterministic, so the fix is to use a regex to
match it.