Clear angular assets manifest
When wanting to avoid openproject cli proxy
(`OPENPROJECT_CLI_PROXY=""`), assets must be precompiled and the
manifest must be built from these assets.

Then if later the assets are deleted like it's done in this script, then
the manifest should reflect that and be cleared too, or it will try to
serve the hashed version of the assets, which is not valid anymore.

This is especially meaningful when running shards with multitenancy
plugin in development mode as the assets are still served by the angular
server, but the routing occurs at a higher level (caddy server doing
reverse proxy). If the manifest exists with hashed values, rails will
return that path with the hash and the angular server cannot serve it.
It needs the unhashed path to serve it.